# Documentation/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DocManSys-RestAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/RestAPI/Document/{id}": {
      "get": {
        "tags": [
          "Document"
        ],
        "summary": "Retrieves a specific document by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the document.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested document.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "If there's an error retrieving the document from the data source."
          },
          "404": {
            "description": "If the document is not found."
          }
        }
      },
      "put": {
        "tags": [
          "Document"
        ],
        "summary": "Updates an existing document by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the document to update.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The updated document object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "If the document was successfully updated."
          },
          "400": {
            "description": "If the provided ID does not match the document ID.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "If there's an error updating the document in the data source."
          }
        }
      },
      "delete": {
        "tags": [
          "Document"
        ],
        "summary": "Deletes a document by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the document to be deleted.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Document successfully deleted."
          },
          "404": {
            "description": "Document not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error while deleting the document."
          },
          "400": {
            "description": "Invalid ID supplied."
          }
        }
      }
    },
    "/api/RestAPI/Document/{id}/upload": {
      "put": {
        "tags": [
          "Document"
        ],
        "summary": "Uploads a file for the specified document and updates the document metadata.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the document to which the file belongs.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "documentFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "documentFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "If the file was successfully uploaded and metadata updated."
          },
          "400": {
            "description": "If the input is invalid (e.g., missing or invalid file type).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "If the document with the specified ID is not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "If an internal server error occurs during the operation."
          }
        }
      }
    },
    "/api/RestAPI/Document/download/{fileName}": {
      "get": {
        "tags": [
          "Document"
        ],
        "summary": "Downloads a file from the server by its name.",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "description": "The name of the file to be downloaded.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File successfully retrieved for download.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "The requested file was not found on the server.",
            "content": {
              "text/plain": {
                "schema": { }
              },
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              }
            }
          },
          "500": {
            "description": "Internal server error while retrieving the file.",
            "content": {
              "text/plain": {
                "schema": { }
              },
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              }
            }
          }
        }
      }
    },
    "/api/RestAPI/Document/search/fuzzy": {
      "post": {
        "tags": [
          "Document"
        ],
        "summary": "Searches for documents using a fuzzy search term.",
        "requestBody": {
          "description": "The term to search for in the documents.",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results successfully returned (list of Documents).",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No matching document"
          },
          "500": {
            "description": "Internal server error while performing the search."
          }
        }
      }
    },
    "/api/RestAPI/Document/search/querystring": {
      "post": {
        "tags": [
          "Document"
        ],
        "summary": "Searches for documents using a search term passed as a query string.",
        "requestBody": {
          "description": "The term to search for in the documents.",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results successfully returned (list of Documents).",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No matching document"
          },
          "500": {
            "description": "Internal server error while performing the search."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "ocrText": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}